[REFACTOR] #153: 리팩토링
