+++ frontend/src/containers/ResultPage/HistogramContainer.tsx
@@ -1,12 +1,17 @@
 import { styled } from 'styled-components';
 import CurveHistogram from '../../components/histogram/CurveHistogram';
 import BarHistogram from '../../components/histogram/BarHistogram';
+import { Dispatch, HTMLAttributes, SetStateAction } from 'react';
 
-export default function HistogramContainer() {
+interface ISimilarQuote extends HTMLAttributes<HTMLDivElement> {
+  setDisplayDimmed: Dispatch<SetStateAction<boolean>>;
+}
+
+export default function HistogramContainer({ setDisplayDimmed }: ISimilarQuote) {
   return (
     <Wrapper>
       <CurveHistogram />
-      <BarHistogram />
+      <BarHistogram setDisplayDimmed={setDisplayDimmed} />
     </Wrapper>
   );
 }
